# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend CMS API untuk SGS Law Firm.\n\nAdmin API menggunakan JWT Bearer Token.\nSebagian besar admin endpoint juga membutuhkan X-Timestamp, X-Nonce, dan X-Signature.\nUpload gambar hanya membutuhkan JWT karena menggunakan multipart/form-data.\n\nFormat canonical X-Signature:\nMETHOD + newline + REQUEST_URI + newline + TIMESTAMP + newline + NONCE + newline + SHA256_BODY",
        "title": "SGS CMS API",
        "contact": {
            "name": "SGS CMS Development Team"
        },
        "license": {
            "name": "Private"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Daftar team admin",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter published status",
                        "name": "published",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerPaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.TeamResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Membuat anggota team",
                "parameters": [
                    {
                        "description": "Team data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/teams/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Detail team admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Memperbarui team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Team data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TeamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "XSignature": []
                    },
                    {
                        "XTimestamp": []
                    },
                    {
                        "XNonce": []
                    }
                ],
                "description": "Melakukan soft delete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Teams"
                ],
                "summary": "Menghapus team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateTeamRequest": {
            "type": "object",
            "required": [
                "name",
                "position"
            ],
            "properties": {
                "biography": {
                    "type": "string",
                    "maxLength": 20000
                },
                "degree": {
                    "type": "string",
                    "maxLength": 100
                },
                "display_order": {
                    "type": "integer",
                    "minimum": 0
                },
                "email": {
                    "type": "string",
                    "maxLength": 150
                },
                "is_published": {
                    "type": "boolean"
                },
                "linkedin_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "photo_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "position": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "practice_area_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "short_description": {
                    "type": "string",
                    "maxLength": 1000
                }
            }
        },
...
        "dto.PracticeAreaResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.SwaggerPaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Data retrieved successfully"
                },
                "meta": {
                    "$ref": "#/definitions/dto.SwaggerPaginationMeta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.SwaggerPaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 25
                },
                "total_pages": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.SwaggerSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.TeamResponse": {
            "type": "object",
            "properties": {
                "biography": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "practice_areas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PracticeAreaResponse"
                    }
                },
                "published_at": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateTeamRequest": {
            "type": "object",
            "required": [
                "name",
                "position"
            ],
            "properties": {
                "biography": {
                    "type": "string",
                    "maxLength": 20000
                },
                "degree": {
                    "type": "string",
                    "maxLength": 100
                },
                "display_order": {
                    "type": "integer",
                    "minimum": 0
                },
                "email": {
                    "type": "string",
                    "maxLength": 150
                },
                "is_published": {
                    "type": "boolean"
                },
                "linkedin_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "photo_url": {
                    "type": "string",
                    "maxLength": 2000
                },
                "position": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "practice_area_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "short_description": {
                    "type": "string",
                    "maxLength": 1000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "XNonce": {
            "description": "Nilai unik untuk mencegah replay attack.",
            "type": "apiKey",
            "name": "X-Nonce",
            "in": "header"
        },
        "XSignature": {
            "description": "HMAC-SHA256 signature dari canonical request.",
            "type": "apiKey",
            "name": "X-Signature",
            "in": "header"
        },
        "XTimestamp": {
            "description": "Unix timestamp dalam detik.",
            "type": "apiKey",
            "name": "X-Timestamp",
            "in": "header"
        }
    }
}